+++ package.json
@@ -15,6 +15,7 @@
     "react-dom": "^18.2.0",
     "react-konva": "^18.2.1",
     "react-scripts": "5.0.1",
+    "recoil": "^0.7.4",
     "typescript": "^4.4.2",
     "web-vitals": "^2.1.0"
   },
