+++ src/atoms/objectAtom.ts
@@ -1,0 +1,7 @@
+import { atom } from 'recoil';
+import { ComponentState } from '../components';
+
+export const componentsAtom = atom({
+  key: 'components',
+  default: [] as ComponentState[],
+});
